+++ src/Cuemon.Extensions.DependencyInjection/ServiceOptions.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using Cuemon.Configuration;
+﻿using Cuemon.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Cuemon.Extensions.DependencyInjection
@@ -8,7 +6,8 @@ namespace Cuemon.Extensions.DependencyInjection
     /// <summary>
     /// Configuration options for Microsoft Dependency Injection.
     /// </summary>
-    public class ServiceOptions : IValidatableParameterObject
+    /// <seealso cref="IParameterObject"/>
+    public class ServiceOptions : IParameterObject
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="ServiceOptions"/> class.
@@ -29,8 +28,6 @@ namespace Cuemon.Extensions.DependencyInjection
         public ServiceOptions()
         {
             Lifetime = ServiceLifetime.Transient;
-            NestedTypeSelector = serviceType => serviceType.GetInterfaces();
-            NestedTypePredicate = _ => true;
         }
 
         /// <summary>
@@ -38,37 +35,5 @@ namespace Cuemon.Extensions.DependencyInjection
         /// </summary>
         /// <value>The lifetime of the service.</value>
         public ServiceLifetime Lifetime { get; set; }
-
-        /// <summary>
-        /// Gets or sets a value indicating whether nested type forwarding should be part of the operation.
-        /// </summary>
-        /// <value><c>true</c> if nested type forwarding should be part of the operation; otherwise, <c>false</c>.</value>
-        public bool UseNestedTypeForwarding { get; set; }
-
-        /// <summary>
-        /// Gets or sets the function delegate that will test each element for a condition based on a <see cref="Type"/>.
-        /// </summary>
-        /// <value>The function delegate that will test each element for a condition based on a <see cref="Type"/>.</value>
-        public Func<Type, bool> NestedTypePredicate { get; set; }
-
-        /// <summary>
-        /// Gets or sets the function delegate that will fetch nested types of a service.
-        /// </summary>
-        /// <value>The function delegate that will fetch nested types of a service.</value>
-        public Func<Type, IEnumerable<Type>> NestedTypeSelector { get; set; }
-
-        /// <summary>
-        /// Determines whether the public read-write properties of this instance are in a valid state.
-        /// </summary>
-        /// <exception cref="InvalidOperationException">
-        /// <see cref="NestedTypePredicate"/> cannot be null - or -
-        /// <see cref="NestedTypeSelector"/> cannot be null.
-        /// </exception>
-        /// <remarks>This method is expected to throw exceptions when one or more conditions fails to be in a valid state.</remarks>
-        public void ValidateOptions()
-        {
-            Validator.ThrowIfObjectInDistress(NestedTypePredicate == null);
-            Validator.ThrowIfObjectInDistress(NestedTypeSelector == null);
-        }
     }
 }